+++ test/test-RenderTest.cc
@@ -41,3 +41,6 @@ TEST_AST("render", "issue-312");
 
 TEST_AST("render", "issue-328-1");
 TEST_AST("render", "issue-328-2");
+
+TEST_AST("render", "fixed-attributes-section");
+TEST_AST("render", "fixed-named-type");